added a single example

+++ Car/include/Car/internal/OpenGL/Texture2D.hpp
@@ -23,7 +23,7 @@ namespace Car {
         virtual uint32_t getHeight() const override { return mHeight; }
         virtual uint32_t getBPP() const override { return mBPP; }
 
-        virtual Rect getRect() const override { return {0, 0, (int32_t)mWidth, (int32_t)mHeight}; }
+        virtual Rect getRect() const override { return {0, 0, (float)mWidth, (float)mHeight}; }
 
         virtual bool operator==(Ref<Texture2D> other) const override {
             return mID == ((OpenGLTexture2D*)other.get())->mID;
